refactor(core): rename test-utils feature and module to testing
