+++ crates/core/src/testing/mem.rs
