+++ crates/node/Cargo.toml
@@ -45,4 +45,4 @@ fs = ["dep:toml", "dep:toml_edit", "dep:ulid"]
 
 [dev-dependencies]
 tempfile.workspace = true
-wcore = { workspace = true, features = ["test-utils"] }
+wcore = { workspace = true, features = ["testing"] }
