+++ crates/runtime/Cargo.toml
@@ -25,4 +25,4 @@ tracing.workspace = true
 
 [dev-dependencies]
 tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
-wcore = { workspace = true, features = ["test-utils"] }
+wcore = { workspace = true, features = ["testing"] }
